Avoid warning
git-svn-id: svn://svn.code.sf.net/p/irrlicht/code/trunk@6350 dfc29bdd-3216-0410-991c-e03cc46cb475

+++ source/Irrlicht/CSTLMeshFileLoader.cpp
@@ -46,7 +46,7 @@ IAnimatedMesh* CSTLMeshFileLoader::createMesh(io::IReadFile* fileIn)
 	if ( fileIn->getType() != io::ERFT_MEMORY_READ_FILE )
 	{
 		u8* fileBuffer = new u8[filesize];
-		if ( fileIn->read(fileBuffer, filesize) != filesize )
+		if ( fileIn->read(fileBuffer, filesize) != (size_t)filesize )
 		{
 			delete[] fileBuffer;
 			return 0;
